Removed an incorrect downgrade of max_size.
This fixes the following situation:
You are using chunked uploads. The total size of the file is 12MB, whereas `min(upload_max_size, post_max_size)` is limited to 8MB. The

The downgrading is still implemented. If you wish to gracefully downgrade max_size, set the configuration value to -1.
This addresses #35.