Include mime type with file upload
When reading the file mime type through php, the information is still in place.  However, when using AWS S3, the uploaded object's metadata always defaults to "application/octet-stream" inside of AWS.  Including the "mimetype" option with the "putstream" function seems to correct this problem; setting the AWS object to the correct content type.